Introduce ability to share ivfpq-l2 table on load
Introduces ability to share model state on index load between segments.
A new cache is introduced, SharedModelStateCacheManager, in order to
properly manage lifecylce of shared model state. Currently, there is no
ability to track the memory taken up by these allocs. But for the
ivfpq-l2 table, it is not currently tracked in the cache anyway.

Signed-off-by: John Mazanec <jmazane@amazon.com>