Add mode and compression based parameter support
Initial commit for mode and compression based parameter support. Everything
compiles but the tests are not in a good state and will need to be revisited in
a later commit, once the dust settles.

Overall, the change includes a major shift towards resolving user
provided parameters to internal objects. This includes allowing
KNNMethodContext and MethodComponentContext to be null as well as
modifying the library logic into a resolution strategy.

For mode and compression, it adds 2 new parameters for either configuring
an index or configuring a model: mode and compression.

After this, the majority of the code that needs to know about the method
configuration needs to go through the field type. More tidying up to
come in the future.

Signed-off-by: John Mazanec <jmazane@amazon.com>