Implement KnnVectorsFormat for engine
Adds a KnnVectorsFormat per engine in a mvp setting. Adds a mapping
parameter called engine that will enable this knnvectorformat to be
used.

Additionally, added a simple query factory extension point on search so
that engines can provide custom queries. Right now, it defaults to
lucene's query.

Was able to validate it works with the following:
```
./gradlew :plugins:bruteforce:integTest
```

Signed-off-by: John Mazanec <jmazane@amazon.com>